[#656] Replaced old dialog by dialog fragment to avoid BadTokenException
