+++ app/src/main/java/org/akvo/flow/ui/fragment/GpsDisabledDialogFragment.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright (C) 2017 Stichting Akvo (Akvo Foundation)
+ *
+ * This file is part of Akvo Flow.
+ *
+ * Akvo Flow is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Akvo Flow is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package org.akvo.flow.ui.fragment;
+
+import android.app.Dialog;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.support.v4.app.DialogFragment;
+import android.support.v7.app.AlertDialog;
+
+import org.akvo.flow.R;
+import org.akvo.flow.ui.Navigator;
+
+public class GpsDisabledDialogFragment extends DialogFragment {
+
+    public static final String GPS_DIALOG_TAG = "gps_dialog";
+
+    private final Navigator navigator = new Navigator();
+
+    public static GpsDisabledDialogFragment newInstance() {
+        GpsDisabledDialogFragment frag = new GpsDisabledDialogFragment();
+        return frag;
+    }
+
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        return new AlertDialog.Builder(getActivity()).setMessage(R.string.geodialog)
+                .setCancelable(true)
+                .setPositiveButton(R.string.okbutton,
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int id) {
+                                navigator.navigateToLocationSettings(getActivity());
+                            }
+                        })
+                .setNegativeButton(R.string.cancelbutton,
+                        new DialogInterface.OnClickListener() {
+                            public void onClick(DialogInterface dialog, int id) {
+                                dialog.cancel();
+                            }
+                        }).create();
+    }
+}
